Expose functional training loop via package entry point
