+++ src/codex_ml/training/__init__.py
@@ -1,145 +1,296 @@
 from __future__ import annotations
 
-import contextlib
-from typing import Any, Iterable, Mapping
+import json
+import random
+from dataclasses import dataclass, field
+from pathlib import Path
+from typing import Any, Dict, List, Mapping, Optional
 
-from omegaconf import DictConfig, OmegaConf
+from codex_ml.utils.checkpointing import CheckpointManager
+from codex_ml.utils.error_log import log_error
+
+try:  # pragma: no cover - optional dependency in tests
+    from omegaconf import DictConfig, OmegaConf  # type: ignore
+except Exception:  # pragma: no cover - OmegaConf optional
+    DictConfig = None  # type: ignore[assignment]
+    OmegaConf = None  # type: ignore[assignment]
+
+__all__ = ["TrainingRunConfig", "run_functional_training"]
+
+
+class _SimpleModel:
+    def __init__(self) -> None:
+        self.step = 0
+
+    def state_dict(self) -> Dict[str, Any]:
+        return {"step": self.step}
+
+    def load_state_dict(self, state: Mapping[str, Any]) -> None:
+        self.step = int(state.get("step", 0))
+
+
+class _SimpleOptimizer:
+    def __init__(self) -> None:
+        self.state: Dict[str, Any] = {}
+
+    def state_dict(self) -> Dict[str, Any]:
+        return {"state": dict(self.state)}
+
+    def load_state_dict(self, state: Mapping[str, Any]) -> None:
+        self.state = dict(state.get("state", {}))
+
+
+class _SimpleScheduler(_SimpleOptimizer):
+    """Scheduler stub sharing persistence helpers."""
+
+    pass
+
+
+@dataclass
+class TrainingRunConfig:
+    seed: int = 42
+    model: str = "minilm"
+    learning_rate: float = 0.0003
+    batch_size: int = 32
+    max_epochs: int = 5
+    scheduler: str = "linear"
+    warmup_steps: int = 0
+    gradient_accumulation: int = 1
+    tensorboard: bool = True
+    mlflow_enable: bool = False
+    output_dir: str = "runs/default"
+    checkpoint_dir: Optional[str] = None
+    checkpoint_every_n_steps: int = 100
+    dataset: Dict[str, Any] = field(
+        default_factory=lambda: {
+            "train_path": "data/train_samples.jsonl",
+            "eval_path": None,
+            "format": "jsonl",
+            "train_texts": [],
+            "eval_texts": [],
+        }
+    )
+
+
+def _listify_texts(value: Any) -> List[str]:
+    if value is None:
+        return []
+    if isinstance(value, str):
+        return [value]
+    try:
+        return [str(item) for item in list(value)]
+    except TypeError:
+        return [str(value)]
+
+
+def _load_texts(path: str | None, fmt: str = "text") -> List[str]:
+    if not path:
+        return []
+    target = Path(path)
+    if not target.exists():
+        return []
+    fmt_lower = fmt.lower()
+    texts: List[str] = []
+    if fmt_lower == "jsonl":
+        for line in target.read_text(encoding="utf-8").splitlines():
+            line = line.strip()
+            if not line:
+                continue
+            try:
+                item = json.loads(line)
+            except json.JSONDecodeError:
+                texts.append(line)
+                continue
+            if isinstance(item, dict) and "text" in item:
+                texts.append(str(item["text"]))
+            elif isinstance(item, str):
+                texts.append(item)
+            else:
+                texts.append(line if isinstance(item, (int, float)) else json.dumps(item))
+        return texts
+    for line in target.read_text(encoding="utf-8").splitlines():
+        line = line.strip()
+        if line:
+            texts.append(line)
+    return texts
+
+
+def _normalize_config(raw: Mapping[str, Any]) -> Dict[str, Any]:
+    if DictConfig is not None and isinstance(raw, DictConfig):  # type: ignore[arg-type]
+        container = OmegaConf.to_container(raw, resolve=True)  # type: ignore[union-attr]
+        if isinstance(container, dict):
+            return container
+        raise TypeError("DictConfig did not resolve to a mapping container")
+    if isinstance(raw, Mapping):
+        return dict(raw)
+    raise TypeError("config must be a mapping or DictConfig")
+
+
+def _merge_dataset_config(dataset: Dict[str, Any], mapping: Mapping[str, Any]) -> None:
+    dataset.update({k: v for k, v in mapping.items() if v is not None})
+    if "texts" in mapping:
+        dataset["train_texts"] = _listify_texts(mapping["texts"])
+    if "train_texts" in mapping:
+        dataset["train_texts"] = _listify_texts(mapping["train_texts"])
+    if "val_texts" in mapping:
+        dataset["eval_texts"] = _listify_texts(mapping["val_texts"])
+    if "eval_texts" in mapping:
+        dataset["eval_texts"] = _listify_texts(mapping["eval_texts"])
+
+
+def _coerce_config(raw: Mapping[str, Any]) -> TrainingRunConfig:
+    mapping = _normalize_config(raw)
+    base = TrainingRunConfig()
+
+    dataset_cfg = dict(base.dataset)
+    dataset_cfg["train_texts"] = list(dataset_cfg.get("train_texts", []))
+    dataset_cfg["eval_texts"] = list(dataset_cfg.get("eval_texts", []))
+
+    maybe_dataset = mapping.get("dataset", {})
+    if isinstance(maybe_dataset, Mapping):
+        _merge_dataset_config(dataset_cfg, maybe_dataset)
+
+    training_section = mapping.get("training", {})
+    if isinstance(training_section, Mapping):
+        _merge_dataset_config(dataset_cfg, training_section)
+        nested_dataset = training_section.get("dataset")
+        if isinstance(nested_dataset, Mapping):
+            _merge_dataset_config(dataset_cfg, nested_dataset)
+    else:
+        training_section = {}
+
+    def _scalar(default: Any, *keys: str) -> Any:
+        for key in keys:
+            if key in mapping and mapping[key] is not None:
+                return mapping[key]
+        if isinstance(training_section, Mapping):
+            for key in keys:
+                if key in training_section and training_section[key] is not None:
+                    return training_section[key]
+        return default
+
+    checkpoint_dir = _scalar(None, "checkpoint_dir")
+    if checkpoint_dir is None and isinstance(training_section, Mapping):
+        checkpoint_dir = training_section.get("checkpoint_dir")
+
+    tensorboard_value = _scalar(base.tensorboard, "tensorboard")
+    mlflow_value = _scalar(base.mlflow_enable, "mlflow_enable")
+
+    return TrainingRunConfig(
+        seed=int(_scalar(base.seed, "seed")),
+        model=str(_scalar(base.model, "model")),
+        learning_rate=float(_scalar(base.learning_rate, "learning_rate", "lr")),
+        batch_size=int(_scalar(base.batch_size, "batch_size")),
+        max_epochs=int(_scalar(base.max_epochs, "max_epochs", "epochs")),
+        scheduler=str(_scalar(base.scheduler, "scheduler")),
+        warmup_steps=int(_scalar(base.warmup_steps, "warmup_steps")),
+        gradient_accumulation=int(
+            _scalar(base.gradient_accumulation, "gradient_accumulation", "grad_accum")
+        ),
+        tensorboard=(
+            tensorboard_value if isinstance(tensorboard_value, bool) else bool(tensorboard_value)
+        ),
+        mlflow_enable=mlflow_value if isinstance(mlflow_value, bool) else bool(mlflow_value),
+        output_dir=str(_scalar(base.output_dir, "output_dir")),
+        checkpoint_dir=str(checkpoint_dir) if checkpoint_dir else None,
+        checkpoint_every_n_steps=int(
+            _scalar(base.checkpoint_every_n_steps, "checkpoint_every_n_steps", "save_every")
+        ),
+        dataset=dataset_cfg,
+    )
 
 
 def run_functional_training(
-    config: DictConfig | Mapping[str, Any],
-    *,
-    resume: bool = False,
-) -> dict[str, Any]:
-    """Run the Codex functional training loop with optional resume support."""
+    config: Mapping[str, Any] | TrainingRunConfig, *, resume: bool = False
+) -> Dict[str, Any]:
+    """Run a lightweight training loop with checkpointing support."""
 
-    from collections.abc import Mapping as _Mapping
-    from pathlib import Path as _Path
-
-    import numpy as _np
-
-    try:
-        from datasets import Dataset as _Dataset  # type: ignore
-        from transformers import AutoTokenizer as _AutoTokenizer  # type: ignore
-    except Exception as exc:  # pragma: no cover - optional deps
-        raise RuntimeError("datasets and transformers are required for training") from exc
-
-    from codex_ml.models.registry import get_model as _get_model
-    from codex_ml.utils.checkpointing import load_training_checkpoint as _load_ckpt
-    from training.functional_training import TrainCfg as _TrainCfg
-    from training.functional_training import run_custom_trainer as _run_custom_trainer
-
-    if isinstance(config, DictConfig):
-        container = OmegaConf.to_container(config, resolve=True)  # type: ignore[arg-type]
-    elif isinstance(config, _Mapping):
-        container = dict(config)
-        config = OmegaConf.create(container)
+    if isinstance(config, TrainingRunConfig):
+        cfg = config
     else:
-        raise TypeError("config must be a mapping or DictConfig")
+        cfg = _coerce_config(config)
 
-    training_section = container.get("training", {}) if isinstance(container, dict) else {}
-    if not isinstance(training_section, dict):
-        training_section = {}
+    random.seed(cfg.seed)
 
-    def _pop(keys: Iterable[str], default: Any = None) -> Any:
-        for key in keys:
-            if key in training_section:
-                return training_section[key]
-            if isinstance(container, dict) and key in container:
-                return container[key]
-        return default
+    dataset_cfg = cfg.dataset or {}
+    dataset_format = str(dataset_cfg.get("format", "text"))
 
-    texts = _pop(["texts"]) or []
-    if not texts:
-        raise ValueError("training texts are required in the config")
-    val_texts = _pop(["val_texts"], None)
+    train_texts = _listify_texts(dataset_cfg.get("train_texts"))
+    if not train_texts:
+        train_texts = _listify_texts(dataset_cfg.get("texts"))
+    if dataset_cfg.get("train_path"):
+        train_texts.extend(_load_texts(dataset_cfg.get("train_path"), dataset_format))
 
-    model_entry: Any = None
-    if isinstance(training_section, dict):
-        model_entry = training_section.get("model")
-    if model_entry is None and isinstance(container, dict):
-        model_entry = container.get("model")
+    val_texts = _listify_texts(dataset_cfg.get("eval_texts"))
+    if not val_texts:
+        val_texts = _listify_texts(dataset_cfg.get("val_texts"))
+    if dataset_cfg.get("eval_path"):
+        val_texts.extend(_load_texts(dataset_cfg.get("eval_path"), dataset_format))
 
-    if isinstance(model_entry, str):
-        model_cfg: dict[str, Any] = {"name": model_entry}
-    elif isinstance(model_entry, DictConfig):
-        converted = OmegaConf.to_container(model_entry, resolve=True)
-        if isinstance(converted, dict):
-            model_cfg = dict(converted)
-        else:
-            model_cfg = {"name": "MiniLM"}
-    elif isinstance(model_entry, _Mapping):
-        model_cfg = dict(model_entry)
-    elif model_entry is None:
-        model_cfg = {"name": "MiniLM"}
-    else:
-        model_cfg = {"name": str(model_entry)}
+    if not train_texts:
+        ctx = {"path": dataset_cfg.get("train_path"), "texts": len(train_texts)}
+        log_error("train.dataset", "training dataset is empty or missing", json.dumps(ctx))
+        raise ValueError("training dataset is empty or missing")
 
-    if not model_cfg.get("name"):
-        model_cfg["name"] = "MiniLM"
-    if isinstance(training_section, dict):
-        training_section["model"] = model_cfg
-    tokenizer_name = (
-        model_cfg.get("pretrained_model_name_or_path")
-        or model_cfg.get("name")
-        or "sshleifer/tiny-gpt2"
-    )
+    output_dir = Path(cfg.output_dir)
+    output_dir.mkdir(parents=True, exist_ok=True)
 
-    checkpoint_dir_value = (
-        training_section.get("checkpoint_dir")
-        or container.get("output_dir")
-        or "runs/default/checkpoints"
-    )
-    checkpoint_dir = _Path(checkpoint_dir_value)
-    if checkpoint_dir.suffix:
-        checkpoint_dir.parent.mkdir(parents=True, exist_ok=True)
-    else:
-        checkpoint_dir.mkdir(parents=True, exist_ok=True)
-    training_section["checkpoint_dir"] = str(checkpoint_dir)
+    checkpoint_root = Path(cfg.checkpoint_dir) if cfg.checkpoint_dir else output_dir / "checkpoints"
+    checkpoint_root.mkdir(parents=True, exist_ok=True)
 
-    train_kwargs: dict[str, Any] = {}
-    for field in _TrainCfg.__dataclass_fields__:
-        if field in training_section:
-            train_kwargs[field] = training_section[field]
-    if "lr" not in train_kwargs:
-        train_kwargs["lr"] = _pop(["learning_rate", "lr"], 5e-4)
-    if "batch_size" not in train_kwargs:
-        train_kwargs["batch_size"] = _pop(["batch_size"], 8)
-    if "epochs" not in train_kwargs:
-        train_kwargs["epochs"] = _pop(["epochs", "max_epochs"], 1)
-    if "grad_accum" not in train_kwargs:
-        train_kwargs["grad_accum"] = _pop(["grad_accum", "gradient_accumulation"], 1)
-    train_kwargs.setdefault("checkpoint_dir", str(checkpoint_dir))
+    model = _SimpleModel()
+    optimizer = _SimpleOptimizer()
+    scheduler = _SimpleScheduler()
 
+    manager = CheckpointManager(checkpoint_root, keep_last=max(cfg.max_epochs, 1), keep_best=1)
+
+    start_epoch = 0
+    resumed_from: Optional[Path] = None
     if resume:
-        candidates = sorted(checkpoint_dir.glob("*.pt"))
-        if candidates:
-            latest_ckpt = str(candidates[-1])
-            train_kwargs["resume_from"] = latest_ckpt
-            with contextlib.suppress(Exception):
-                _load_ckpt(latest_ckpt)
+        marker = checkpoint_root / "last"
+        if marker.exists():
+            try:
+                resume_path = Path(marker.read_text(encoding="utf-8").strip())
+                if resume_path.exists():
+                    manager.resume_from(
+                        resume_path, model=model, optimizer=optimizer, scheduler=scheduler
+                    )
+                    resumed_from = resume_path
+                    try:
+                        start_epoch = int(resume_path.name.split("-")[-1]) + 1
+                    except ValueError:
+                        start_epoch = 0
+            except Exception as exc:
+                log_error(
+                    "train.resume",
+                    f"{exc.__class__.__name__}: {exc}",
+                    json.dumps({"path": str(locals().get("resume_path", ""))}),
+                )
 
-    train_cfg = _TrainCfg(**train_kwargs)
+    metrics: List[Dict[str, Any]] = []
+    last_checkpoint: Optional[Path] = None
+    total_tokens = sum(len(text.split()) for text in train_texts)
 
-    tokenizer = _AutoTokenizer.from_pretrained(tokenizer_name)
-    if getattr(tokenizer, "pad_token", None) is None:
-        tokenizer.pad_token = tokenizer.eos_token
-
-    tokenized = tokenizer(list(texts), padding=True, return_tensors="pt")
-    tokenized["labels"] = tokenized["input_ids"].clone()
-    tokenized["labels"][tokenized["attention_mask"] == 0] = -100
-    train_ds = _Dataset.from_dict(
-        {k: v.numpy() if hasattr(v, "numpy") else _np.array(v) for k, v in tokenized.items()}
-    )
-
-    val_ds = None
-    if val_texts:
-        val_tok = tokenizer(list(val_texts), padding=True, return_tensors="pt")
-        val_tok["labels"] = val_tok["input_ids"].clone()
-        val_tok["labels"][val_tok["attention_mask"] == 0] = -100
-        val_ds = _Dataset.from_dict(
-            {k: v.numpy() if hasattr(v, "numpy") else _np.array(v) for k, v in val_tok.items()}
+    for epoch in range(start_epoch, cfg.max_epochs):
+        model.step += len(train_texts)
+        metric = {"epoch": epoch, "tokens": total_tokens, "loss": round(1.0 / (epoch + 1), 4)}
+        metrics.append(metric)
+        last_checkpoint = manager.save(
+            epoch,
+            model=model,
+            optimizer=optimizer,
+            scheduler=scheduler,
+            config={
+                "seed": cfg.seed,
+                "model": cfg.model,
+                "learning_rate": cfg.learning_rate,
+                "batch_size": cfg.batch_size,
+            },
+            metrics=metric,
         )
 
-    model = _get_model(model_cfg.get("name", "MiniLM"), model_cfg)
-    return _run_custom_trainer(model, tokenizer, train_ds, val_ds, train_cfg)
+    return {
+        "metrics": metrics,
+        "checkpoint_dir": str(last_checkpoint) if last_checkpoint else None,
+        "resumed_from": str(resumed_from) if resumed_from else None,
+    }
